Create separate VM for FavouriteLocationsScreen.kt
