+++ app/build.gradle
@@ -89,6 +89,7 @@ dependencies {
 
     //Hilt
     implementation "com.google.dagger:hilt-android:$hiltVersion"
+    implementation "androidx.hilt:hilt-navigation-compose:$androidXHiltVersion"
     kapt "com.google.dagger:hilt-compiler:$hiltVersion"
 
     // Datastore
