--- app/src/main/java/com/example/weatherapp/feature/favouritelocations/util/FavouriteLocationsUIState.kt
@@ -1,9 +1,0 @@
-package com.example.weatherapp.feature.favouritelocations.util
-
-import com.example.weatherapp.feature.favouritelocations.model.ShortWeatherInfo
-
-sealed class FavouriteLocationsUIState {
-    object Error : FavouriteLocationsUIState()
-    object Loading : FavouriteLocationsUIState()
-    data class Success(val data: List<ShortWeatherInfo>) : FavouriteLocationsUIState()
-}
