+++ build.gradle
@@ -16,6 +16,7 @@ buildscript {
         junitVersion = "4.13.2"
         okHttpVersion = "4.10.0"
         hiltVersion = "2.44.2"
+        androidXHiltVersion = "1.0.0"
         protobufVersion = '3.21.7'
         dataStoreVersion = '1.0.0'
     }
